Separate out indexer test
